+++ test.js
@@ -1,0 +1,58 @@
+
+
+const a = true; const b = true;
+const c = function () { return; };
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+
+if ( a  &&  b )
+  c();
+  
+if( a  &&  b )
+  c();
+
+  
+if ( a  ||  b )
+  c();
+  
+if( a  ||  b )
+  c();
+
+
+
+
+
+
+
+
+